09-key-data-extraction/pyproject.toml
[tool.poetry]
name = "09-key-data-extraction"
version = "0.1.0"
description = ""
authors = ["julio colomer <info@aiaccelera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
jupyterlab = "^4.2.4"
langchain = "^0.3.7"
langchain-core = "^0.3.20"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.1.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
